third_party/blink/web_tests/virtual/css-anchor-update-enabled/external/wpt/css/css-anchor-position/position-area-scrolling-006-expected.txt

This is a testharness.js-based test.
[FAIL] Scroll to 194
  assert_equals: anchored y expected 194 but got 1200
Harness: the test ran to completion.
